feature-policy/reporting/wakelock-reporting.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

const t = async_test("WakeLock Report Format");

const check_report_format = (reports, observer) => {
  let report = reports[0];
  assert_equals(report.type, "feature-policy-violation");
  assert_equals(report.url, document.location.href);
  assert_equals(report.body.featureId, "wake-lock");
  assert_equals(report.body.sourceFile, document.location.href);
  assert_equals(typeof report.body.lineNumber, "number");
  assert_equals(typeof report.body.columnNumber, "number");
  assert_equals(report.body.disposition, "enforce");
};

new ReportingObserver(t.step_func_done(check_report_format),
                      {types: ['feature-policy-violation']}).observe();

t.step_func(() => {
  assert_throws('NotAllowedError', () => new WakeLock("screen"),
      "WakeLock API should not be allowed in this document.");
})();
</script>
